Fix workflow generic node port name
