+++ ee/codegen/src/context/node-context/base.ts
@@ -1,5 +1,3 @@
-import { isEmpty, isNil } from "lodash";
-
 import { VELLUM_WORKFLOW_NODES_MODULE_PATH } from "src/constants";
 import { WorkflowContext } from "src/context";
 import { PortContext } from "src/context/port-context";
@@ -129,36 +127,12 @@ export abstract class BaseNodeContext<T extends WorkflowDataNode> {
     return toPythonSafeSnakeCase(nodeOutputName, "output");
   }
 
-  private isPortNameUsed(portName: string): boolean {
+  public isPortNameUsed(portName: string): boolean {
     return this.nodePortNames.has(portName);
   }
 
-  private addUsedPortName(portName: string): void {
+  public addUsedPortName(portName: string): void {
     this.nodePortNames.add(portName);
-  }
-
-  public generateSanitizedPortName(portName: string): string {
-    const defaultName = "port_";
-    const rawPortName = portName;
-
-    const initialPortName =
-      !isNil(rawPortName) && !isEmpty(rawPortName)
-        ? toPythonSafeSnakeCase(rawPortName, "port")
-        : defaultName;
-
-    // Deduplicate the port variable name if it's already in use
-    let sanitizedName = initialPortName;
-    let numRenameAttempts = 0;
-    while (this.isPortNameUsed(sanitizedName)) {
-      sanitizedName = `${initialPortName}${
-        initialPortName.endsWith("_") ? "" : "_"
-      }${numRenameAttempts + 1}`;
-      numRenameAttempts += 1;
-    }
-
-    this.addUsedPortName(sanitizedName);
-
-    return sanitizedName;
   }
 
   /**
